Massive update to the base site, configuration, CSS and templates.
Did a bundle update, made sure to vendor the gems inside the directory.

Made a bunch of changes to CSS, especially code highlighting, font size
and choice. I updated the MathJax version to use a CDN that has not been
deprecated (since 2017, yikes!). I also udpated from FontAwesome 4.5,
which was vendored, to using a kit tied to my account.